refactor(convert): extract border from extract_block_style
